pyproject.toml
[tool.poetry]
name = "ppo-gymnasium-lunarlander"
version = "0.1.0"
description = ""
authors = ["Giansimone Perrino <hello@giansimone.dev>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
numpy = "^2.3.4"
torch = "^2.9.0"
gymnasium = {extras = ["box2d"], version = "^1.2.2"}
pyyaml = "^6.0.3"
imageio = {extras = ["ffmpeg"], version = "^2.37.2"}
huggingface-hub = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
